+++ mm2src/mm2_bitcoin/spv_validation/src/lib.rs
@@ -6,9 +6,6 @@ extern crate serialization;
 extern crate sha2;
 extern crate test_helpers;
 
-/// `types` exposes simple types for on-chain evaluation of SPV proofs
-pub mod types;
-
 /// `helpers_validation` Override function modules from bitcoin_spv and adapt for our mm2_bitcoin library
 pub mod helpers_validation;
